crewai-hello/pyproject.toml
[project]
name = "crewai-hello"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[litellm]>=1.14.1",
    "python-dotenv>=1.1.1",
]
